Adds enum validators seeds and connects through validation to auth. Renames a mis-named seedfile.
