+++ src/db/seedFiles/20_enum_validators.js
@@ -1,0 +1,8 @@
+// deno-lint-ignore-file
+const validators = [{ validator: 'user' }, { validator: 'parent' }];
+
+async function seed_validators_enum(knex) {
+  await knex('enum_validators').insert(validators);
+}
+
+module.exports = { seed_validators_enum };
